Move Dialog code to DialogBase and built Dialog on top of it
