+++ src/components/dialog/index.js
@@ -1,4 +1,5 @@
 import Dialog from './Dialog';
+import DialogBase from './DialogBase';
 
+export { Dialog, DialogBase };
 export default Dialog;
-export { Dialog };
